+++ src/main/java/com/CliftonBoyd2007/AudioBug/core/AudioBug_Init.java
@@ -44,7 +44,7 @@ public final class AudioBug_Init implements EditorFactoryListener {
     private void audioBugStartupNotifier(FeedbackService feedbackService) {
         if (!this.notifiedUserOfStartup) {
             feedbackService.player.playCue_Start();
-            feedbackService.announce("AudioBug is ready.", true);
+            feedbackService.announce("AudioBug started.", true);
             this.notifiedUserOfStartup = true;
         } else {
             return;